created service for reporting
